Add player class
Initialize class with an integer id, a class variable containing the
player count and a mark variable containing the symbol that the player
will use to mark the board.

Seperation of methods is such that all methods that need to act on a
specific player are in the Player class (e.g. checking if a specific
player has won). This will prevent (or at least minimize) passing around
Player objects as arguments.