Add Game class and main script
Add tic_tac_toe.rb as main entry point for the game which also contains
all of the requires to make it easier to manage these.

Add the Game class which will for now be the only class to contain any
output. These output methods might later be moved to their own seperate
module. Also all game constants like winning conditions will be put in
the Game class as well.